Finalizando a edição de dados de usuario

+++ controllers/LoginControllers.js
@@ -12,15 +12,17 @@ const controller = {
         let erros = validationResult(req);
 
         if(!erros.isEmpty()){
-            fs.unlink(path.resolve(__dirname + '/../public/img/profile-img/' + req.file.filename), () => console.log(`Arquivo ${req.file.filename} excluído...`));
+            if(req.file !== undefined){
+                fs.unlink(path.resolve(__dirname + '/../public/img/profile-img/' + req.file.filename), () => console.log(`Arquivo ${req.file.filename} excluído...`));
+            };
             res.send('Erro ' + JSON.stringify(erros.errors));
         }else{
             let {nome_usuario, cpf_usuario, email_usuario, senha_usuario, telefone_usuario, data_nasc_usuario} = req.body;
-            let foto_usuario = '/img/profile-img/' + req.file.filename;
+            let foto_usuario = req.file === undefined ? 'null' : '/img/profile-img/' + req.file.filename;
             let senhaEnc = bcrypt.hashSync(senha_usuario, 10);
 
             let usuario = await Usuario.create({nome_usuario, cpf_usuario, email_usuario, senha_usuario: senhaEnc, telefone_usuario, data_nasc_usuario, foto_usuario});
-            console.log(usuario);
+            // console.log(usuario);
             
             res.send('Usuario Criado com sucesso!');
         };
@@ -30,9 +32,6 @@ const controller = {
 
         if(usuario.length > 0 && bcrypt.compareSync(req.body.senha_usuario, usuario[0].dataValues.senha_usuario)){
             req.session.usuario = usuario[0].dataValues;
-            console.log(req.session.usuario);
-            console.log('-----------------');
-            console.log(usuario[0].dataValues);
             res.redirect('/');
         } else{
             res.send('email ou senha errados, verifique os dados e tente novamente');
